Fix for some parcel media not loading properly (DEV-39135).
There were a couple of issues interacting to cause this:

Firstly, when LLViewerMediaImpl delayed loading the parcel media (which it normally does), it was losing track of the specified MIME type, but the autodiscovery code also wasn't being triggered.  The code should now carry through both the specified MIME type and the autodiscovery flag when loading is delayed.

Second, the new media autodiscovery code might not work for some legacy parcel media content (for example, if it's stored on a server that doesn't report the correct MIME types).  The code has been changed to first check whether the specified MIME type maps to a known plugin and allow that to override MIME type detection, which should allow this legacy content to keep working.