DEV-40210 : restructure media controls UI
This change completely restructures the media controls XUI
so that it is now more flexible so as to implement mini vs.
standard controls.   It also now allows for changes based
on artwork and design still as unyet delivered from 80/20,
as well as any feedback from Brian's user testing.

This is also step 1 of renaming the media "hud" to media "controls"