EXT-2429 body parts out of sync with COF
Visual param definitions weren't working properly when the wearable being
loaded had fewer visual parameters than the current definition of the wearable.
This occurred when you were loading wearables that had been created in older
versions of the client (or from the library).

Added missing parameters to mSavedVisualParamMap, and modified revertValues()
to ensure that the two maps are kept fully in sync on finishing loading the
wearable and when reverting the values.

Code reviewed by Vir