EXT-2539 confirm we're not sending local texture IDs in appearance msg
created a throwaway function to temporarily set local texture IDs in the avatar
object to IMG_DEFAULT_AVATAR immediately before sending the TE message, and
restore the IDs when we're done. Will implement a more elegant solution for 2.1.
This is the least risky implementation, though it isn't elegant.

Code reviewed by Seraph