Fix for DEV-43040 (Change media focus rules to be the same as alt-zoom).
Added an optional reset_camera parameter to LLPanelPrimMediaControls ::resetZoomLevel(). Default is true, but if false it prevents the reset from moving the camera, and just resets the internal state for zoom tracking.

Fixed zoom-related bugs in LLPanelPrimMediaControls::close() and LLPanelPrimMediaControls::onClickZoom().

Changed the zooming behavior in LLViewerMediaFocus to match the desired behavior.