DEV-43439: Allow Home side panel URL to be customized.
The URL for the Home side panel is now specified via a new
"HomeSidePanelURL" string in app_settings/settings.xml. This
string supports the following substitutions:

CHANNEL = the channel name for the viewer
VERSION = the full version string for the viewer
LANGUAGE = the current language set in the viewer
AUTH_KEY = the authentication key (see below)

The authentication key is an optional string that is read from the
dictionary of strings returned by login.cgi. If login.cgi returns a
key called "home_sidetray_token", then the value of that key is used
for the AUTH_KEY substitution. This lets the server provide a piece of
blind data that can be passed to the Home panel web page to support
authentication. The viewer does not interpret the token in any way.

This change adds a new module, llviewerhome.{cpp|h} to contain the
model functionality, used by the view module, llpanelhome.{cpp|h}.