EXT-3394: Remove manual voice gain controls.
The new Vivox SDK supports Automatic Gain Control (AGC), so we
shouldn't need to change the local gain level from the default of 50%.

This change sets the initial gain level to 50 (instead of 62) and
removes the local gain sliders from the Sound Preferences floater and
the Voice Controls floater. The latter was not hooked up yet anyway.