DEV-44838 - This is an attempt at the minimal possible change to fix the quoting bug on windows.
This commit is untested!  The simplest approach to testing is to push into a repo watched by parabuild.

Also, develop.py could be improved to use subprocess consistently across all platforms.  Doing so could simplify the code, but until I understand how to test it better, I'm going to leave it alone.