EXT-5480: Support secondlife:///app/region SLapps.
This is defined to work exactly like the classic secondlife://{REGION}
Slurl. However, the later syntax cannot support spaces in the region
name because spaces (and %20 chars) are illegal in the hostname of an
http URL. Some browsers let you get away with this, but some do not
(such as Qt's Webkit). Hence we have decided to introduce the newer
secondlife:///app/region alternative.