AVP-92 FIX Avatar appears to be fat when opening the shape editor
Wearable editor previews were affecting the in-world avatar and not properly
resetting the changed parameters. Since the "body fat" slider is the last in
the default tab for the shape editor, this leaves your avatar fat.

Reconfigured the code to work through the wearables and most importantly
included a call to updateVisualParameters() at the end of the render function.

This appears to have fixed the issue - your avatar does not appear affected during
the avatar preview generation anymore.


Code reviewed by Seraph