EXT-8213 FIX users cannot replace their shape if it does not load
Removed checking for wearables loaded on replacing or adding individual items.
After reviewing the code in depth, we believe this is safe to do,
particularly since we allow the user to replace their outfit from the same state.

Filed a followup issue for later investigation EXT-8231

Code reviewed by Vir and Seraph