EXT-8244 FIXED Not wearable links could be added to COF by drag-n-drop in the Inventory
Fixed the dnd behavior so that you can only drag appropriate items into the COF or Outfits folder.
Did a bit of superficial code cleanup in terms of adding comments and relocating blocks of code.  The actual coding fix was pretty small despite the diff; most of the diff is superficial cleanup.