EXT-8342 FIXED Fixed displacement of decription textbox in Place Profile after reshape.
Bug was reproducible on Windows and not on Linux because of different moment when onTopLost() was called for textbox.
On Windows it happened only after reshape, on Linux- before reshape. Made windows behaviour in case of top lost consistent with Linux behaviour.

- An if was added into LLPanelPlaceInfo::reshape to force collapsing description textbox on Windows at the beginning of reshape
(the only case when reshape is skipped there now is when it's caused by expanding/collapsing this textbox, so called_from_parent for panel is FALSE).

Reviewed by Richard Nelson at https://codereview.productengine.com/secondlife/r/779