VWR-22052 FIX tons of processAvatarAppearance spam, may be flooding server
Added a timeout so that we will only send this message a max of once a minute.
Investigated cause of server not sending visual params with texture message,
it appears it only happens if the sim cannot find the visual params to send.
Not sure what causes that condition.

Code reviewed by Seraph