STORM-153 FIXED Localized Conference and non-existent IM session name strings.
- Moved the "no session" string to XML and used it in code instead of hardcode.



- Conference string turned out to be already translated, but the way translation was happening ad-hoc chat log files on disk
would get localized names, and thus be inconsistent (and that would create problems with browsing them later).
It didn't happen already just because this string wasn't translated in other locales somewhy.
So moved localization of session name after the the creation of log file.
Also changed value of string according to Eli's comment in ticket.