ER-219: Add an asynchronous api for the region debug console
Add support for a new capability, SimConsoleAsync. This is a replacement for
the deprecated SimConsole capability that provides an asynchronous API. The old
API is still supported, but the simulator will never support both.

For the new API, commands will be sent via HTTP POST to the SimConsoleAsync
capability, same as before. But the resulting output will be pushed to the
viewer view the event-poll, using the newly added SimConsoleResponse LLSD
message.

Will review with Simon or Falcon.