EXP-772 -- Log in failure, keeps saying DNS cannot resolve hostname.
No real progress on this Jira yet but Mac build was not properly reporting the CURL error string.
This check-in fixes that.

So far, I have backed out URL related changes between 2.6.2 and 2.6.3 without any change in behavior.
Unsure how to proceed next although comparing libcares and libcurl builds between 2.6.2 and 2.6.3 seems
like it could be the next logical step.

Users experiencing the problem can go back to vewere 2.6.2 or adjust DNS settings to use the google free
DNS servers 8.8.8.8 and 8.8.4.4 which should allow them to work around the problem.

Reviewed by Richard.