EXP-859 FIX -- Save setting for last time inbox expanded EXP-841 PROGRESS -- Create outbox sync button with basic enable/disable logic and animation EXP-866 PROGRESS -- Create inbox observer to dynamically update inbox item count and badge value EXP-872 PROGRESS -- Hide inbox and outbox unless applicable
Inbox expansion time is now saved as setting variable "LastInventoryInboxExpand".

New outbox sync button art.

Observers for items and categories are now implemented in the LLSidepanelInventory but the exact
method of use is still to be determined.

Functions are in place to show and hide the inbox and outbox panels but the logic around the
decision is TBD.

The resize bar for the inbox/outbox is now disabled when both inbox and outbox are collapsed.

Reviewed by Leyla.