STORM-1186 FIXED Removed the "no mCurrentRMessageTemplate" warning message.
The mCurrentRMessageTemplate member seems to only be non-NULL for a short while
after an incoming message was validated and parsed, thus there is no
guarantee that we can obtain name of the last received message at any given
time. So if we can't we'll simply return an empty string without spamming the
log with warnings.