STORM-1620 FIXED Re-enabled saving settings.
Reason: settings.xml was mistakenly marked as a per session (not saved)
settings file. See description of the "SessionSettingsFile" setting.

Fix: It looks like the proper way to make the Advanced mode default
is to leave the setting value empty, which is what this change does.
It also re-fixes STORM-1616.