SH-4034 FIX Outfit changes broadcast without exiting appearance edit mode
Backing out MAINT-1099 fix. Added line causes us to temporarily exit and re-enter
outfit editing mode every time we edit a wearable, which has a host of unintended
consequences. Reviewed the original bug, and their fix should not have any
impact on the original bug.

Repro of the original bug does not repro, with or without the patch. The patch
causes regressions but not fixes so removing it. Will send MAINT-1099 back
to QA for verification - there is an alternate solution if this causes a regression.